PowerShell and Automation Hyper-V Health Check Tool
Recently, I was engaged in an environment where SCOM 2012 R2 has been in use for all monitoring activities; including monitoring of Hyper-V host and VMs. But, SCOM itself was a VM which was residing inside the physical Hyper-V host. Thus, if and when the physical Hyper-V host goes down, SCOM and monitoring also goes down. Hence, I thought of writing this Hyper-V Health-Check tool which can be executed from remote machines as well as from the physical hyper-V host.The Hyper-V Health Check Tool – HyperVHCTool.exe – offers both GUI mode as well as the command line parameters for execution. The latter is particularly user-friendly in the instance of executing the health check tool as a schedule task which can recurring take place, as good operational and administrative practice, as per schedule as defined by the administrator. This way, the need for automation is also achieved. Creating a schedule task to run twice everyday addresses the need to automatically perform health-check of the Hyper-V hosts.

This tool supports both local as well as remote execution. This tool can be used directly on the physical Hyper-V host machine to perform health-checks; either via the GUI mode or by using the command line parameters. Or, the administrator can execute this tool from his local desktop/laptop or any other machine and use the Hyper-V hosts as the target servers on which health-checks are to be performed.  In either of the above, it must be enabled and ensured that the target Hyper-V hosts must be reached and contacted over the network using RPC/WMI. For optimum results, this tool must be executed by Admin users who have administrative privilege in the Windows Active Directory domain.