PowerShell: File and Folder Search Tool With GUI
I like the way I can search files and folders in good old Windows 2003 Servers. It offered a nice GUI from the Control Panel. With the latest operating systems, that old GUI is no longer available. I don't mean to say that the present GUI options are not upto my taste; but its just that I stil admire the good old GUI that was offerred in Windows 2003 servers.

Hence, I wrote this tool and I wish to share it with you all. This tool enables the User to search for any file or folder in the computer; including hidden files and folders. Just type the file or folder name to search and check the results.

How To
=======
Extract all contents of the zip file to a folder; e.g. C:\Tools\SearchTool
Do not delete any item from the extracted zip file
Right-Click the EXE and select "Run As Administrator"

Your feedback and comments are welcome. Please feel free to rate your experience.